+++ src/components/Choice.tsx
@@ -1,21 +1,9 @@
 import { ItemAction, SessionError } from '@dialob/fill-api';
 import { useFillActions, useFillValueSet, useFillSession } from '@dialob/fill-react';
 import React, {useMemo} from 'react';
-import { Select, MenuItem, InputLabel, FormControl } from '@material-ui/core';
-
-import { Theme } from '@material-ui/core/styles';
-import { makeStyles, createStyles} from '@material-ui/styles';
-
+import { Select, MenuItem, InputLabel, FormControl } from '@mui/material';
 import { ErrorHelperText } from './helpers';
 import { DescriptionWrapper } from './DescriptionWrapper';
-
-
-const useStyles = makeStyles((_theme: Theme) => createStyles({
-  formControl: {
-    minWidth: 120,
-  },
-})
-);
 
 export interface ChoiceProps {
   choice: ItemAction<'list'>['item'];
@@ -23,7 +11,6 @@ export interface ChoiceProps {
 };
 
 export const Choice: React.FC<ChoiceProps> = ({ choice, errors }) => {
-  const classes = useStyles();
   const session = useFillSession();
   const {setAnswer} = useFillActions();
   const valueSet = useFillValueSet(choice.valueSetId);
@@ -41,8 +28,8 @@ export const Choice: React.FC<ChoiceProps> = ({ choice, errors }) => {
 
   return (
     <DescriptionWrapper text={choice.description} title={choice.label}>
-    <FormControl fullWidth={true} required={choice.required} error={errors.length > 0} className={classes.formControl}>
-      <InputLabel id={`${itemId}_label`}>{choice.label}</InputLabel>
+    <FormControl fullWidth={true} required={choice.required} error={errors.length > 0} sx={{minWidth: 120}}>
+      <InputLabel id={`${itemId}_label`} shrink>{choice.label}</InputLabel>
       <Select labelId={`${itemId}_label`}
         value={choice.value || ''}
         onChange={e => setAnswer(choice.id, e.target.value)}