+++ src/components/Group.tsx
@@ -1,17 +1,8 @@
 import { ItemAction } from '@dialob/fill-api';
 import React, {useContext} from 'react';
-import { Paper, Typography, Grid, Fade } from '@material-ui/core';
-import { Theme } from '@material-ui/core/styles';
-import { makeStyles, createStyles } from '@material-ui/styles';
+import { Paper, Typography, Grid, Fade } from '@mui/material';
 import { GroupContext } from '../context/GroupContext';
 import { Description } from './Description';
-
-const useStyles = makeStyles((theme: Theme) => createStyles({
-  paper: {
-    padding: theme.spacing(2)
-  }
-})
-);
 
 type ColumnType = boolean | "auto" | 2 | 1 | 12 | 6 | 3 | 4 | 5 | 7 | 8 | 9 | 10 | 11 | undefined;
 
@@ -20,7 +11,6 @@ export interface GroupProps {
 };
 
 export const Group: React.FC<GroupProps> = ({ group, children }) => {
-  const classes = useStyles();
   const groupCtx = useContext(GroupContext);
   let columns = group.props?.columns || 1;
   columns = columns > 4 ? 4 : columns;
@@ -31,7 +21,7 @@ export const Group: React.FC<GroupProps> = ({ group, children }) => {
   return (
     <GroupContext.Provider value={{level: groupCtx.level < 6 ? groupCtx.level + 1 : groupCtx.level}}>
       <Fade in={true}>
-        <Paper elevation={groupCtx.level} className={classes.paper}>
+        <Paper elevation={groupCtx.level} sx={{p: 2}}>
           <Grid container spacing={2}>
             <Grid item xs={12}>
               <Typography variant='h3'>