+++ src/components/RowGroup.tsx
@@ -1,19 +1,11 @@
 import { ItemAction } from '@dialob/fill-api';
 import { useFillActions } from '@dialob/fill-react';
 import React from 'react';
-import { Typography, Grid, Button, } from '@material-ui/core';
-import { Theme } from '@material-ui/core/styles';
-import { makeStyles } from '@material-ui/styles';
-import { Add } from '@material-ui/icons';
+import { Typography, Grid, Button, } from '@mui/material';
+import { Add } from '@mui/icons-material';
 import { FormattedMessage } from 'react-intl';
 import { Description } from './Description';
 import { RowGroupContext } from '../context/RowGroupContext';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  addButton: {
-   float: 'right'
-  }
-}));
 
 export interface RowGroupProps {
   rowGroup: ItemAction<'rowgroup'>['item'];
@@ -21,14 +13,13 @@ export interface RowGroupProps {
 
 export const RowGroup: React.FC<RowGroupProps> = ({ rowGroup, children }) => {
   const {addRowToGroup} = useFillActions();
-  const classes = useStyles();
   return (
     <>
       <Grid item xs={12} style={{ marginBottom: '5px' }}>
         <Typography variant='h3'>
           {rowGroup.label || <span>&nbsp;</span>}
           <Description title={rowGroup.label} text={rowGroup.description} />
-          <Button size='small' color='primary' variant='contained' className={classes.addButton} onClick={() => addRowToGroup(rowGroup.id)} startIcon={<Add />} style={{ marginBottom: '3px' }}><FormattedMessage id='row.add.button' /></Button>
+          <Button size='small' color='primary' variant='contained' sx={{float: 'right'}} onClick={() => addRowToGroup(rowGroup.id)} startIcon={<Add />} style={{ marginBottom: '3px' }}><FormattedMessage id='row.add.button' /></Button>
         </Typography>
       </Grid>
       <RowGroupContext.Provider value={{rowGroup}}>