+++ src/intl/en.ts
@@ -16,7 +16,9 @@ const en = {
   'row.remove.button': 'Remove',
   'row.remove.confirmation.title': 'Do you want to remove the row?',
   'row.remove.confirmation.cancel': 'Cancel',
-  'row.remove.confirmation.confirm': 'Remove'
+  'row.remove.confirmation.confirm': 'Remove',
+
+  'autocomplete.nooptions': 'No options'
 };
 
 export default en;